refactor: split notion connector modules
